backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.4
billiard==4.2.4
cachetools==6.2.2
celery==5.6.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
diff-match-patch==20241021
Django==6.0.1
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-import-export==4.3.9
django-querycount==0.8.3
django-unfold==0.64.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
kombu==5.6.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==7.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.9.0
